requirements.txt
python-telegram-bot==20.8
stripe
requests
python-dotenv
